start ce
