+++ deployments/src/commands/report.py
@@ -1,0 +1,46 @@
+"""Report command - generate test reports."""
+
+from pathlib import Path
+from typing import Optional
+
+import typer
+from typing_extensions import Annotated
+
+from ._shared import check_initialized, console
+
+
+def report(
+    deployment_id: Annotated[
+        Optional[str],
+        typer.Argument(help="Deployment ID to generate report for (optional)")
+    ] = None,
+    output: Annotated[
+        Optional[Path],
+        typer.Option("--output", "-o", help="Output file path (default: .arm-testing/results/)")
+    ] = None,
+    format: Annotated[
+        str,
+        typer.Option("--format", "-f", help="Report format (json/yaml/markdown)")
+    ] = "markdown",
+) -> None:
+    """
+    Generate test report for deployments.
+
+    If no deployment_id specified, generates a summary report of all deployments.
+
+    Examples:
+        uv run neo4j-deploy report
+        uv run neo4j-deploy report abc123
+        uv run neo4j-deploy report --format json --output report.json
+    """
+    check_initialized()
+    console.print("[yellow]Report command not yet implemented[/yellow]")
+
+    if deployment_id:
+        console.print(f"[cyan]Would generate report for:[/cyan] {deployment_id}")
+    else:
+        console.print(f"[cyan]Would generate summary report for all deployments[/cyan]")
+
+    console.print(f"[cyan]Format:[/cyan] {format}")
+    if output:
+        console.print(f"[cyan]Output:[/cyan] {output}")
